Rename cmd_entries subdir to cmd
