+++ sourced/ml/__main__.py
@@ -6,14 +6,14 @@ import sys
 from modelforge.logs import setup_logging
 
 from sourced.ml.extractors import IdentifierDistance
-from sourced.ml.cmd_entries import bigartm2asdf_entry, dump_model, projector_entry, bow2vw_entry, \
+from sourced.ml.cmd import bigartm2asdf_entry, dump_model, projector_entry, bow2vw_entry, \
     run_swivel, postprocess_id2vec, preprocess_id2vec, repos2coocc_entry, repos2df_entry, \
     repos2ids_entry, repos2bow_entry, repos2roles_and_ids_entry, repos2id_distance_entry, \
     repos2id_sequence_entry
-from sourced.ml.cmd_entries.args import add_df_args, add_feature_args, add_split_stem_arg, \
+from sourced.ml.cmd.args import add_df_args, add_feature_args, add_split_stem_arg, \
     add_vocabulary_size_arg, add_repo2_args, add_bow_args, add_repartitioner_arg, \
     ArgumentDefaultsHelpFormatterNoNone
-from sourced.ml.cmd_entries.run_swivel import mirror_tf_args
+from sourced.ml.cmd.run_swivel import mirror_tf_args
 from sourced.ml.utils import install_bigartm
