+++ sourced/ml/tests/test_id2vec.py
@@ -8,7 +8,7 @@ import numpy
 import sourced.ml.tests.models as paths
 from sourced.ml.models import Id2Vec
 from sourced.ml.utils import projector
-from sourced.ml.cmd_entries import projector_entry
+from sourced.ml.cmd import projector_entry
 
 
 class Id2VecTests(unittest.TestCase):
